flask-crud/requirements.txt
flask
flask-cors
azure-storage-blob
python-dotenv
psycopg2-binary
postgres